requirements.txt
requests
pyhamcrest
xmltodict
pytest
mock
